# Optimized numpy functions in Converter