This PR optimized the conversion speed.
This is done mainly by choosing more performant variants for the numpy processing in the `Converter` class.

The Converter now also gets whole batches from the `Predict.queue_out_frames` method. This was mainly done to enable batch processing and potential replacing numpy functions with keras ones later on, but it also increased my speed by about 3it/s.

Optimizing the worker Plugins remains to be done. They did not take enough time in my setup to be really relevant. This may be different with other settings.

Benchmarks:
- Current staging: 16.45 it/s
- This PR WITHOUT providing whole batches to the converter: 21.40 it/s
- This PR as is: 24.69 it/s

Tested on 10k frames with a resolution of 1280x720.
The tested model on my system can do raw prediction (without anything else) with about 37.2 it/s