@kilroythethird 2019-08-15
Added kvroomans suggestions.
Should be good to go